[Fix apache/incubator-kie-issues#987] Refactor ProcessNodeEvent children
Although not strictly needed by the PR, for future additions, a new
class in the hierarchy that holds the NodeInstance reference has been
included.
fields has been made final to remark that event objects are inmutable
